# backend/docs/swagger.yaml
basePath: /api/v1
...
info:
  contact: {}
  description: API do sistema de ressarcimento (SURE)
  title: SURE API
  version: "1.0"
paths:
...
  /api/v1/fluxo-ressarcimento/{id}:
    get:
      parameters:
      - description: ID do processo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Busca fluxo de ressarcimento
      tags:
      - FluxoRessarcimento
    post:
      consumes:
      - application/json
      parameters:
      - description: ID do processo
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Salva fluxo de ressarcimento
      tags:
      - FluxoRessarcimento
  /api/v1/fluxo-ressarcimento/item/{itemId}:
    delete:
      parameters:
      - description: ID do item
        in: path
        name: itemId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Deleta item do fluxo de ressarcimento
      tags:
      - FluxoRessarcimento
...
  /api/v1/tags:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Lista tags
      tags:
      - Tags
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Cria tag
      tags:
      - Tags
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"